requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-grappelli==4.0.1
django-jet==1.0.8
django-recurrence==1.11.1
django-suit==0.2.28
djangorestframework==3.15.2
drf-spectacular==0.28.0
fastapi==0.115.7
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
tzdata